// libs/reports/common/src/lib/enums/report-fields.enum.ts
export enum ReportFieldDeviceTypeIdEnum {
    Desktop = 0,
    Smartphone = 1,
    Tablet = 2,
    FeaturePhone = 3,
    Console = 4,
    Tv = 5,
    CarBrowser = 6,
    SmartDisplay = 7,
    Camera = 8,
    PortableMediaPlayer = 9,
    Phablet = 10
}

export enum ReportFieldDeviceTypeNameEnum {
    Desktop = 'desktop',
    Smartphone = 'smartphone',
    Tablet = 'tablet',
    FeaturePhone = 'feature_phone',
    Console = 'console',
    Tv = 'tv',
    CarBrowser = 'car_browser',
    SmartDisplay = 'smart_display',
    Camera = 'camera',
    PortableMediaPlayer = 'portable_media_player',
    Phablet = 'phablet'
}